Module Dolmen_intf

module Ext : sig ... end

Interfaces for Extensions

module Id : sig ... end

Interfaces for Identifiers. This module defines Interfaces that implementation of identifiers must respect in order to be used to instantiated the corresponding language classes.

module Language : sig ... end

Interface for Dolmen languages modules

module Lex : sig ... end

Interface for Dolmen lexers.

module Location : sig ... end

Interface for file locations. This module defines an interface to store locations in files. These locations are used by the parsers to specify the locations of all parsed expressions whenever it is possible. This interface also requires some exceptions to be defined. These exceptions make use of locations to specify at which point in the file the parsing went wrong.

module Map : sig ... end

module Msg : sig ... end

Messages

module Parse : sig ... end

Interface for Dolmen parsers.

module Pretty : sig ... end

Pretty printing annotations

module Stmt : sig ... end

Interfaces for statements. This module defines interfaces for statements, i.e top-level declarations in files.

module Tag : sig ... end

Interfaces for Tags. Tags are used to attach arbitrary information to ndoes in an ast.

module Term : sig ... end

Interfaces for Terms. This module defines Interfaces that implementation of terms must respect in order to be used to instantiated the corresponding language classes.

module Tok : sig ... end

Interface for token descriptions This doe snot define an interface for token descriptions but rather defines a type for token description, that will then be used by others interfaces.

module Ty : sig ... end

Interfaces for Types This module defines Interfaces that implementation of types must respect in order to be used to instantiate functors.