Module Dolmen_intf.Id

Interfaces for Identifiers. This module defines Interfaces that implementation of identifiers must respect in order to be used to instantiated the corresponding language classes.

Signature for Logic languages

module type Logic = sig ... end

module type Response = Logic

module type Escape = sig ... end