Module type Term.Smtlib_Array

type t

The type of terms

type ty

Type of of types.

val const : ty -> t -> t

const ty base creates a constant array that maps any value of type ty to base.

val select : t -> t -> t

select arr idx creates the get operation on functionnal array arr for index idx.

val store : t -> t -> t -> t

store arr idx value creates the set operation on functional array arr for value value at index idx.