Module type Ty.Smtlib_Float

Signature required by types for typing smtlib bitvectors

type t

The type of types

val bitv : int -> t

Create a fixed size bitvector type.

val float : int -> int -> t

Create a float type with fixed exponent size in bits and fixed significand, including the hidden bit.

val roundingMode : t

Type of the rounding modes

type view = private [>

| `Real

| `Bitv of int

| `Float of int * int

]

Partial views for types. These are used in the Float theory to perform type-base dispatch for some conversion functions.

val view : t -> view

Partial view of a type.